replace Q with native Promises
This effort is stopped. Since we have to support Node 6, and Util.Promisify is only added in Node 8, we needs Q.nfcall. Also we use the handy Q.spread.
It makes no sense to remove the dependency on Q, and write our own code to replace these, as long as we need to support Node 6.